Added tests for Login
The tests run on the basis of the nextjs page (and NOT the storybook
component).

The nextjs page had to be changed because the storybook component relies
on the complete LoginViewModel and this model is only queried after the
initial render. I am not sure whether it is the best practice to have an
if/else clause that only renders the storybook component if the
LoginViewModel is not undefined, but it is what it is for now.

The tests concern themselves with the collapsed divs depending on the
input given by the LoginViewModel, which is passed to nextjs by virtue
of fetchMocks.