Added mock flask endpoint for rseaccountusage
Simply copied the code for route `/stream` and changed for
`/rseaccountusage`, the data conforms to
`core/data/rucio-dto.ts::RSEAccountUsageLimitDTO` which conforms to the
output given by the actual rucio server.

Note that the flask server code is not really nice (because it is
redundant, things i could have abstracted away still remain) but i dont
think that this is an issue.