Select Row Works
Selecting a row works in the new RSEQuotaTable (streamed). Selection
works exclusively via then checkbox to the right, but I want to add
tapping the row as well.

Tanstack seems to be working against me. I hoped to use the tanstack API
to save the selected spaces and then use `useEffect` linked to the
tanstack table (or, even better, the selected rows within the table),
but that just would not work. Hence, pressing the checkbox will update
the react state and then update the tanstack state.

I hope to find a better solution soon.