Spinnywheel on table head if query is fetching
Theres a clash with the cursor-pointer setting, ie if you hover over a
row then there will be a pointer and no spinnywheel, and I am not sure
if this is desired behaviour or not.

Will definitely need to set a minimum size for the table though so empty
space will show a spinnywheel during backgrond stream fetches.