Streaming Table is now Scrollable
The best solution to have a dedicated space which is empty and with a
border that gets filled with table rows (while keeping the table header
at the top when scrolling) is not the same way as I did it earlier.

Instead of overflowing `tbody`, the entire `table` is placed into an
overflowing `div`. The sole `tr` of `thead` is `sticky` and `top-0` in
tailwindcss terms.

The border between head and the `tbody` is provided
by a shadow (since borders are somehow not visible in the sticky setup).
The shadow gives a nice effect.