Selection of RSE Disabled when no quota, Switch to `Remaining Quota`.
TODO: Add the `ask permission` functionality, but this should not be an
issue thanks to the `isNoQuotaLeftFunction` function where this can be
set for the entire page.

~~

Until now, it was possible to select RSEs with no quota left by tapping
the row (the checkbox was already disabled), this is now no longer
possible.

Furthermore, we used to show `Used` and `Total` quota, I have changed
this to `Remaining` and `Total` Quota because this makes more sense to
humans (as opposed to the previous setting, which is simply what Rucio
puts out).