Created Story for (Streamed) RSE Quota Table
The Streamed RSE Quota Table can now be found among the components under
`StreamedTables/RSE Quota Table`.

The issue was that the ComDOM-hook has several requirements that I do
not fully understand. Note the `QueryClientProvider` under
`src/app/rseaccountusage/layout.tsx` which wraps the table. This is not
good for stories, which should be completely encapsulated.

Hence, the two statuses that are required to properly display the table
(`fetchstatus` and `data`) were extracted as parameters. `data` is a
JSON Array with `RSEAccountUsageLimitDTO` objects as elements.

At this stage, there is no utility to return the selection to the parent
element (or vice versa: no way to tell the table which elements to show
as selected).