Added `DIDListTable` StreamedTable
Currently only shows the DID Name and the Size (plus a column for the
checkboxes).

The assumption at the moment is that a query is only done for one
specific scope, this means that the DID **NAME** is unique -- if the
query were to be run on more than one scope, this would no longer be the
case, as the scopenames are unique throughout rucio, not the names.