Added tags
Added responsive tags (pills) that can show the DID Type. Doing it this
way instead of using a column is at once a sensible design choice as
well as an exercise in programming.

The pills are responsive and will resize to circles with only the first
letter (*C*ollection, *D*ataset, *F*ile). Collections have been
completely omitted, even though they are formally a part of the
`DIDType` type.

Colour choice:

Collection => Teal
Dataset => Amber
File => Rose

Dark mode is already implemented.

Also allowed selection of the full DID-Name upon clicking the field -- I
have the feeling that this is a good idea in principle, but the
execution is not good. Might need to be rolled back.