Dropdown Menu for Page Size, Fixed Page (No Overflow/Scrolling)
On the DIDListTable.

Up to this point, the table had a fixed size regardless of the number of
rows to be shown per page, meaning that the table was paginated AND
scrolling. This has been changed to a fixed size (depending only on the
number of rows to display).

I am not sure what the best move is here... 100 rows per page is so much
that even people on large screens will need to scroll.