Added Sorting to RSEQuotatable for Quota Columns
I wonder what the best move is about the icon to use when the column is
not sorted. I opted to use the HeroIcons (NOT HeroIcons2) triple
horizontal dots `HiDotsHorizontal` because it indicates that something
can be clicked. I think it might be a bit misleading but it is much
better than having no indication at all. I also added a cursor-pointer
to show the user that this element can be clicked.

The added benefit to using the HeroIcons is that they are MIT-licensed
and that I would stay within the icon set that is recommended for
tailwind and that I have begun using earlier.

Also changed the column sizing:
* Small widths will have a hidden total quota `quota_bytes` column, with
  the name and remaining quota `remaining_bytes` columns having the same
  width.
* Above small widths, the quota columns will have `w-1/3` and the name
  column has width `w-1/2`, in practice the flex-engine will apply
  different ratios.
* In both cases, the `selection` column will be present and have `w-8`,
  not growing or shrinking at all.