Added Input to Filter Results of RSEName column
You can now use a text filter for the RSE Name.

For normal-width viewports, the input form will be to the right of the
column head text within the header.

For smaller widths, the input will collapse into a magnifying glass
icon. Clicking the icon will replace the entire table head with a search
form. A checkmark icon to the right will return the table head row to
its original state, but keep the filter state.

Dark mode is also implemented.

Known issue:
since the input fields for the small viewport and the larger ones are
separate HTML entities, their values are not linked by default, and I
have not made any change to that. This means that unexpected behaviour
can come up if one enters text into the input field for small viewports,
and then switches back to the larger viewport (without clearing the
input field first). => the value of the larger viewport input will not
reflect the entry into the smaller field, while the content of the table
will remain filtered.