Sort by Type
It is possible to filter by type. Pressing the head cell of the `Type`
column will cycle between

`undefined`
`File`
`Dataset`
`Container`

in that order. `Undefined` allows any DIDType and the icon is shown to
be the triple horizontal dots, in line with the other sorts. The icon
will then change to the `DIDTypeTag` in its new `forcesmall` form.

Known Issues:
* The icon of the DIDTypeTag is only fully rounded when below
  `medium`-sized viewports (even with `forcesmall` option), since the
  rounding is handled in tailwind.
* The DID-Name is not necessarily related to its type in the way the
  mock server is currently set up... this means that
  `container-jflcn...` is actually a File. But this is just the
  cosmetics of the mock python server.