Fixed Type Error in useRef
Remember that `useRef(null)` doesnt just initialise to a value `null`,
but will remember that type. It follows that when you try to access the
`.current`-attribute, the linter will expect this to be `null`.

Using `useRef<HTMLDivElement>(null)` in this case allows the linter to
accept that the `useRef` targets a `<div>`-element or null. Handle the
remaining error by using the `?.`-syntax.

https://www.designcise.com/web/tutorial/how-to-fix-object-is-possibly-null-typescript-error-when-using-useref-react-hook