Fixed bug that tapping the account icon will reopen the dropdown
clicking the button for the accounts toggles the `isProfileOpen`
state.

Previously, tapping outside the account dropdown would close the
dropdown. This meant that clicking the button would first close and then
toggle the dropdown, meaning that it would remain open albeit for a
short flicker.

In this fix, I added the button reference to the clicking logic, so that
clicking the button has no secondary effect on closing the dropdown --
it only acts as a toggle. This means that clicking the button when the
dropdown is open actually closes the dropdown, instead of the confusing
earlier effect.