Fixed React issue `ReferenceError: window is not defined`
Which was caused by referring to the `window.inner{Width, Height}`
without actually having a `window`. This is not an issue when the user
renders the page, because in that case we have a browser and hence also
a window. But react also renders the page on the server, where no window
is defined.

I have already fixed this for `DIDSelectTable`