Used Spread Syntax; Close Tab on next Elem
Tabbing to the next element after the dropdown will close the search
dropdown. Tabbing backwards will not close the dropdown.

Bug: Tabbing to the right does not immediately create a focus ring
around the first link. Why is this the case?

Achieved this effect by using the spread syntax, meaning that the
elements take the parameters in the form of `...props` with props being
of type `JSX.IntrinsicElements["a"]`..