Rework Button Class for Disabled Cursor
Hovering over a disabled button will now give an disabled cursor, called
`cursor-not-allowed` in Tailwind.

In order to do so, the component was reworked to make more use of
`twMerge`, instead of using the outdated array concatenation.

The signature of the component is completely unchanged and the
appearance of the login page for example is the same.