Reworked Dropdown to Pure `Grouping by DIDType` Dropdown for `CreateRule`
In the spirit of creating only the necessary components and being as
bottom-heavy as possible, the generic dropdown component has now been
reworked to exist solely as an item grouping dropdown for the
`CreateRule` page.

Includes `DIDTypeTags` in the button text (reponsive here) and in the
dropdown (will always remain fully-sized here).

The label in `CreateRule` has been changed back to a simple pure
`<label>` that is hidden and only exists for screen-readers.