Added `SamplingTag`
Design in teal. Changed the `DIDTypeTag`-colour for the Container from
teal to emerald.

The difference between the `SamplingTag` and the other tags at the
moment is that the text is bold and in black/white (whereas the other
tags are not bold and the text colour is a darker/lighter version of the
bg-colour).