Reflect differences in didtype in `rucio-server.py`
As per my conversation with Mario, I expect to receive a fixed number of
key-value pairs (with known keys) for the didmetadata. There are minor
differences between the metadatas for files and colections.

I only have one API endpoint (here: `/didmeta`) that returns data of the
same format irrespective of the datatype. But keys will have
None-values when they should not be set for the given datatype.

The further handling will need to be done on the webserver.