Mocked the DIDMetaQuery
Wanted to use a stream and then tanstack, decided against it. Reason is
that I only have one row (because we only query metadata for one DID),
so streaming rows is not necessary. In addition, this would mean working
*against* the framework, which I do not want to do.

We expect the backend to provide the `ListDID` component with a single,
complete object that can then be parsed by the `DIDMetaView` component.
A single boolean parameter can be used to communicate to the components
whether the data is ready to be rendered.