Added subpages and first state to PageDID
`PageDID` now has a state storing which subpage is shown. The situation
is made more complex because different subpages are visible depending on
whether the DID is a collection or a file:

File Replica States / Parent DIDs if file
Rules / Dataset Replicas / File Replica States if collection

in that order.

I also changed the `Tabs` component to use `data-id` as an attribute,
this is accessed from typescript via `dataset.id`. Keep in mind that
`data-*` attributes can be used  to assign arbitrary data to an HTML
element and can always be accessed in typescript via `dataset.*` (thanks
Chat-GPT).