Added `ReplicaStateTag`
with the type `ReplicaState` added to `core/entity/rucio.ts`, it
reflects the enum defined in `rucio.db.sqla.constants::ReplicaState`.

Each state has an associated colour:
* `Available`/green
* `Unavailable`/red
* `Copying`/amber
* `Being_Deleted` (printed as "Being Deleted", ie with space)/gray
* `Bad`/pink
* `Temporary_Unavailable` (printed as "Temp. Unavailable" to save
  space)/blue

The design of the tag is as follows:
* width chosen to fit the widest state string ("Temp. Unavailable")
* Black text, italic (white in dark mode)
* Background in 200 (700 for dark mode)
* Border in 700 (200 for dark mode)

At the moment, the tag is not responsive.

The tag text can be selected regularly. I think this makes more sense
than making it unselectable. In particular, one can select and then copy
an entire table, including the tags, as if they were text.