Added `PageDIDFilereplicasD` for Datasets
New table which only has a single `DID` column, containing the DIDs of
files contained within a dataset. Selecting a single file is possible
and will show all the replica states in a table shown below (this is the
same table as in `PageDIDFilereplicas`, used as a component.

Filtering by name is possible.

The table expects the DID to be provided as `scope` and `name`
separately. In addition, the number of replicas in each of the
ReplicaStates (eg Available/Unavailable/Copying/...) must be provided.
This is in order to add tags later. *It can also be taken out, should
this be a large query*

Placed a large modified `File` `DIDFileTag` on top of the selectable DID
table.