`PageDIDRules`
In doing so, also added `RuleStateTag`

Each state has an associated colour:
* `OK`/green
* `Stuck`/red
* `Replicating`/amber
* `Suspended`/gray
* `Waiting_Approval` (printed as "Waiting Approval")/pink
* `Inject`/blue

The design of the tag is as follows:
* width chosen to fit the widest state string ("Waiting Approval")
* Black text, underlined (white in dark mode)
* Background in 400 (700 for dark mode)
* Border in 700 (200 for dark mode)

The design is very similar to the `ReplicaStateTag`, in particular, the
colours are the same (with the background colour intensity in light mode
having been bumped up from 200 to 400). I do not think that this is an
issue. I am running out of obvious colours after all, and the
differentiation needs to be done via hues and text decorations (I am
avoiding icons for now).

Filtering the columns is possible.