`PageDIDContents` ==> Use new column defs
Almost identical to `PageDIDParents`, I decided to use a different
component anyway because I do not want issues in the future when one
table begins to differ more from the other.

`DIDType` column can be hidden in `PageDIDContents`

I can now use the tanstack-provided utility to hide table columns. In
order to do so, I had to put the headers into the columndef as well,
which is difficult when the `th`-tag itself is required to have style
properties.

The issue was solved by adding a `meta` object attribute and packing the
`style` string into it, and then accessing this string and passing it
into the auto-generated `th`-tags.

In order to avoid typing errors, a new `StyleMetaColumnDef<T>` type was
created and stored in `streamedtables.d.ts`, this extends a generic
`ColumnDef<T>` by adding the `meta` object with the `style` attribute.

see: https://github.com/TanStack/table/discussions/4179

Added dark mode to `PageDIDContents`

Added `PageDIDContents` to `PageDID`