`PageDIDDatasetReplicas`
Also added `RSETag`:

which augments the simple RSE Name by also adding the icons whether the
RSE is blocked from read/write/delete (or a combination of those).

This information is optionally passed in via an integer between 0 and 7.

Issue that came up: (`FullFilter`)
Fixed input losing focus issue on the new `PageDID` subpages

The new `PageDID`-subpages `PageDIDContents` and
`PageDIDDatasetReplicas` make use of a new way of describing the table,
i.e. moving not only the cell definitions but also the header
definitions into the `columns` constant (array). This worked well,
except for the fact that the `Filter` component lost focus after
entering a single character, rendering the text input useless.

This solution extends the `meta` object (as described in
`StyleMetaColumnDef<T>`) by an optional `filter: boolean` attribute. If
this attribute is set and positive, the new `FullFilter` component will
be rendered instead of the custom header cell defined in the `columns`
array constant.

`FullFilter` wraps the pre-existing `Filter` component even more,
meaning that it renders into a complete cell, including the responsive
design for mobile, in which the input box collapses into a magnifying
glass. Since `FullFilter` also handles the `windowSize` and the derived
boolean of `smallScreenNameFiltering` itself, it is no longer necessary
to have this code in the parent file.