**Create generic StreamedTable**
* Pass in tableData and an array of columnDefs (which I have not
  successfully managed to type yet) and it will render a table.
* Also has `tableStyling`, which currently only supports the
  `columnVisibility`.
* Filtering by discrete values is possible.
* Filtering by string values is possible (also for small screens, copied
  behaviour from small screens: textbox collapses into a magnifying
  glass button).
* Pagination via the `PaginationDiv` component. This div is placed
  within the `tfoot` component of the table, this means that all
  components of the table are contained in one `table` tag (instead of
  having an overall parent `div` tag).
* Has `FetchStatusIndicator`
* Component for `TableExternalLink` (based on the `a` tag but styled
  like buttons).
* Possible to choose column width in the `meta.style` object of the
  column definition. This is still pretty buggy, since tailwind does not
  recognise any classes defined only here and will not compile them. The
  styles will only be applied if the styles were defined elsewhere as
  well. #TODO!
* Begun work on adding table fixtures.

**Notes**
Filtering capability Notes:

* do not use any additional state management, just `get` and `set` using
  the functions provided by `info.column.{get, set}FilterValue()` =>
  this is IMPORTANT!
* use an object to store the next filter entry (so you can cycle
  through). Remember that the default unset value for tanstack is
  `undefined`.
* cast the value of `getFilterValue()` to `keyof typeof nextRecord`
  where `nextRecord` is the object containing the next filter value in
  order to avoid type errors

Added table fixture For `StreamedTable`, which uses `DIDContents` for
test purposes. Will extend the fixtures later.

Added table to locks of `PageRule` using the parametrised table!

Added `PaginationDiv` into the table, solved issue by putting it in
`tfoot`, thus keeping the entire StreamedTable component within a
`table` tag. Make the table cell have
`colSpan={table.getVisibleLeafColumns().length}`

Added `tableStyling` option, currently only supports `columnVisibility`:
`columnVisibility` is a setting passed as a `Record<string, bool>`. The
user can define whether (and under which conditions) a column shall be
visible.

Pass column width into generic table

The generic table is of `table-layout: fixed` (tailwind: `table-fixed`).
This means that the width of a column is determined by the width of the
cell in the first row, i.e. the header.

Pass the style for the header (and hence also the width) to the generic
table via the `meta` parameter. It is included in the extended
`ColumnDef` which I called `StyleMetaColumnDef<T>`.

**Installed packages**:
Installed `faker-js` package