Added `TableFilterDiscrete` story
This component is added into the header of a column. It contains both
the header name and the discrete filtering icon, as well as the logic
cycling through the filtering options and connecting to the table.

Parameters:
* `name` is just the string that is passed to the `H3` tag
* `keys` is an array of type `T`, usually the strings of an enum. They
  indicate which states the filter can take.
* `renderFunc` takes a key or `undefined` and shall return a
  `JSX.Element`, this will likely be an icon. `undefined` means that no
  filter value is set (the default tanstack value)
* `column` the tanstack table column of type `Column<any, T>`, whose
  `setFilterValue` will be called.

Added `TableFilterDiscrete` to `PageRule`

Created `TableFilterString` and implemented

in `PageRule`.

Added `TableFilterString` support for small widths

Remember that the input field will give way to a magnifying glass
button. Clicking it will create an input field the full width of the
header row, which will collapse upon clicking a checkmark.

Responsiveness of `PageRule` Locks table

Changed `PaginationDiv` to only require `table` as argument

We no longer handle the table pagination state in a separate react state
parallel to the tanstack table state.

Cleanup of `PageRule`, added dark mode.