`PageRule` test, assert tabbing through subpages works
Note that testing a full tab runthrough of the page is not possible to
the best of my knowledge. This is because tabbing will jump to items
which should normally be hidden by the responsive design, but the
testing library does not properly handle responsive design.

Example: in normal-width displays, tabbing from `Filter DID` textbox
should lead to the `Filter RSE` textbox in the Locks table of
`PageRule`. However, the test will jump to the magnifying glass button
that would only be visible in small width displays instead.

TODO: write more tests!