**Custom Extraction Logic**
In a `StreamedTable`, the styles for a column header `<th>` set the
width of the entire table column, because the `StreamedTable` has class
`table-fixed`. In order to pass arbitrary styles (strings) into the
`StreamedTable`, one can add an object to `meta` of a
`columnHelper.accessor` or `columnHelper.display`:

```
// ... id, header def and cell def...
meta: {
	style: ""
}
```

This has been typed in `StyleMetaColumnDef<T>`, which extends a
`ColumnDef<T>` from tanstack.

The issue was that tailwind did not recognise these style strings during
compile-time. We added custom extraction logic (see [Tailwind
Docs](https://tailwindcss.com/docs/content-configuration#customizing-extraction-logic)).

The regex will match any valid style string defined under the `style`
attribute of the `meta` object. Note that `style` needs to be the first
element of `meta` for the regex to work.