Added `Filter` and `Replication Rules` as `Code` to `PageSubscription`
Added syntax highlighting to `Code` component

`Code` will now:
  * Handle light/dark mode using `useEffect` and `useState`
  * use the `Highlight` component (with the language set to
    `javascript`), the component is imported from `react-hightlight`

In order to add the custom styles, the `Highlight` component has been
wrapped in an `html` and `body` tag. This means that a `Head` tag (from
`nextjs` containing a `style` tag can be used to add the themes as
strings.

Installed `@types/react-highlight` and `react-highlight` modules.