Added improved schema validation
Checked the schemas from the transmogrifier and cast them into
`TypeBox`-schemas to be used in the website. They can now be found under
`core/entity/subscription.ts`, I decided to place them *outside* of the
general `rucio.ts` file because of the special import (of `Type` from
`@sinclair/tybebox`) they need.

The schemas can also be converted to strings to be displayed within the
`PageSubscriptionJSONEditor`. Press the help-button (questionmark-icon,
also with an ARIA-compliant label) to view the string version of the
schema.