Pass Selection State to Parent
At the moment, the data is passed from `StreamedTable` to the direct
parent via `StreamedTableProps.tableselecting?.handleChange`. This
function should be a `useCallback` from the parent.

The data passed is an array of objects of datatype `T` (i.e. the
datatype to which the streamedtable has been parametrised). We do not
work with individual keys but with the object as a whole. This is a
possible area of improvement.

(This is supposed to be a generic solution, so we do not know a priori
which key could uniquely identify the table rows to its parent
components).