Updated table in `PageDIDFilereplicasD`
We need to make more use of the datatype `FileReplicastateD`, which has
much more information than just the scopename that simply isnt being
displayed properly.

The top table can currently be single-selected. No logic exists to
handle the selection state in the parent component, this will need to be
implemented later.