Worked on tables in `CreateRule`
This means adding:
* `CreateRuleDIDTable`
* `CreateRuleRSETable`

The following old verisons were removed:
* `RSEQuotaTable` (ancient version, non-streaming -> removed `Tables`
  folder)
* `RSEQuotaTable` (streaming version)
* `DIDSelectTable`

Removing the two old streaming tables also means that the
`StreamedTables` folder is now mostly clean of legacy files. The goal is
to exclusively have `StreamedTable` and associated components inside
that folder, no standalone-tabes.

Added the `helpers.tsx` file for the StreamedTables, which currently
only contains the helper function `didToScopename`, which converts an
array of `DID` (or any extensions of `DID`, such as `DIDLong`) into an
array of DID scopenames.

Also made changes to:

* `TableSortUpDown`: added option to `nocollapse`, meaning that the
  string will not hide at widths below "md"
* `Number` was changed to only show `NaN` for real `NaN` and handle
  negative numbers properly.
* Added fixture for `RSEAccountUsageLimit`

Note regarding `CreateRule` state:

Since the selection state is no longer stored as an array of keys but as
an array of original data objects (thus better mirroring the internal
data state of the tanstack table), the `Page0State` has been changed: it
is no longer an array of DID scopenames but an array of `DIDLong`
objects (containing scope, name, etc.)

**Note regarding nextjs**

the nextjs pages do not work properly. Will deal with this later.