Rework StreamedTable: Nav Elements and ComDOM Query
* created `tablefooter`
* changed `paginationdiv` to `paginationnav` by name
* replaced old footers
* added control buttons to `TableFetchStatus`
* Added TableFooter to StreamedTable
* added Breakout
* Removed FetchStatusIndicator
* Removed obsolete `Filter` component

Also Connected ComDOM pause-resume

Issue: I want to be able to initialise the ComDOM with a query, so that
`.start()` can be called without arguments

Start ComDOM Query from within Component Page

Until now, the ComDOM was initialised and the query started from within
the nextjs page. This meant that it was not possible to start or alter
the query from a lower level of the table stack (`nextjs > component
page > StreamedTable`). In essence, changing or starting the query in
response to user interaction was not possible, since user interaction is
handled in `component page` and `StreamedTable` and was not propagated
back into the nextjs page.

With the ComDOM now passed by reference into the lower levels and the
exposed `request` and `setRequest` attributes of `UseComDOM<T>`, it is
now possible to start the ComDOM-query of `PageRule` once the `Locks`
tab is opened. So far, the query was started on page load. Now, the user
has to open the tab showing the streamed table for the query to run.

The query will not run if the table is already populated. The user can
manually rerun the query using the UI provided by `TableFetchstatus`