Added `tableFooterStack` styling option to `StreamedTable`
for pages such as `ListDID`, in which one third of horizontal space is
reserved for a metadata block, it the `TableFetchstatus` component might
intersect with the `TablePagination` component. In order to avoid this,
the new styling option `tableFooterStack: boolean` optionally stacks the
components vertically even on medium and large width layouts.