Worked on basic widget `WidgetOngoingrules`
Is a widget that displays the ongoing rules (up to 20 of them) in a
stacked bar chart. The input is an array of `Ongoingrule` objects, which
contain the rulename as well as the number of locks in each state.
Internal logic parses these numbers and returns a stacked bar chart
using percentages. The colours correspond to those used in the light
mode `LockStateTag`.

Added `test/fixtures/widget-fixtures.ts`, since widget fixtures are
distinct from tables.

For the same reason, added
`lib/infrastructure/data/view-model/widgets.d.ts` to take widget
datatypes.