Built `WidgetUsedquota` widget
Will show 6 (3 on small viewports) panels for RSEs in a grid layout. The
used and free space are shown in a pie chart. Also shows a table with
relevant and formatted information about the quota.

Information shown:
* RSE Name
* Space used
* Quota -> this is not directly shown in the chart. The chart shows
  *free* space, which is `quota - space used`
* Total space -> this is not shown in the chart, only in table, it is
  additional information
* May exceed quota (boolean, shown in a `BoolTag`) -> only in table

On small screens, the widget shows one column of three panels with only
the RSE name and the chart (no legend). The legend and the table appear
on wider screens.

Components using `chart.js` will not be fully responsive, i.e.
increasing the screen width will lead to weird behaviour as shrunk
components will not resize to the larger screen. In practice, this
should not be an issue for now. *Should deal with later, though*