Replaced `Date` type by `DateISO = string`
Added new type `DateISO = string` to `rucio.ts`. Since JSON does not
support `Date` objects, the values will be parsed as ISO strings. @maany
and I decided to handle the conversion back to dates in the front-end.

Hence, this commit changes the exported types, the table fixtures and
the calls to the data (which now convert strins aka `DateISO`s back into
dates).

Removed `Date` type from `PageDIDMetadata`

just handled as a string now, no detection of strings that are formatted
as strings.