Added Tanstack Table Backend to WidgetOngoingRules
Added sorting to the widget!

Used the tanstack table backend to feed into the widgets, generalised
this. Using tanstack means that we can use a framework that is already
known to us in order to implement sorting, filtering, etc.

When creating a new widget, bear in mind to use `ChartJS` only for very
specific purposes (i.e. the chart itself) and implement other UI
elements, such as sorting and filtering using either custom components
or those taken from the `StreamedTable`.