New page: `List RSE`, added `useResponsiveHook`
LIST_RSE: Added types, fixtures

Renamed old `RSE` type to `RSEOld` (couldnt delete because used by
@maany), because we need different attributes:

* dropped `city`, `country`, `continent`, `latitude`, `longitude`
* added `deterministic` and `staging_area` booleans
* changed `rse_type` to new enum `RSEType` which can be tape or disk.

Added types and fixtures.

Added streamedtable for `ListRSE`

added `stack` option to `TableFilterDiscrete`

Added `TableFilterBoolean`

the current `TableFilterDiscrete` was not compatible with booleans, so I
added a filter component. It is in use in `ListRSE`.

Wrapped StreamedTable with a page in `ListRSE`

Added `TableInternalLink` and used it in `ListRSE`

Added `components/Helpers/ResponsiveHook.ts`

This is a custom react hook `useResponsiveHook` returning an object of
type `ResponsiveHook` which is currently a bunch of booleans
corresponding to the width breakpoints of tailwind. Use in conjunction
with `StreamedTable` to simplify the syntax:

`StreamedTable` takes the `tablestyling` option, which has the
`visibility` object. The keys of this object are the column ids, and the
values are the booleans determining whether the column is shown.

**e.g.:**

for a given column `xyz`, the column should be visible above the
tailwind medium breakpoint. hence, pass

```
visibility: {
  xyz: responsive.md
}
```
into `StreamedTable`s `tablestyling`. You need to initialise the hook
with `const responsive = useResponsiveHook()` earlier. This will emulate
the same behaviour as `className="hidden md:block"` could (not directly
possible in the case of table columns).

**Possible extension**
`StreamedTable` needs to be able to handle vertical window height as
well, in order to choose the appropriate amount of rows to show. Extend
this hook by writing handlers for the vertical window height.

**Note** that we use this hook *outside* of `StreamedTable` and pass the
results *into* the table because there can be other React components in
the page that might need to react responsively to window dimension
changes. It would not make sense to call this hook twice, once in the
parent and once in the child.

Finished `ListRSE`