Added `Modal` for Login Multiaccounts select
Within the login UI, users can select an account, or have one mapped to
them. If multiple accounts can be mapped to the given credentials, the
new modal will show up, prompting the user to select an account.

To facilitate this, added the optional `rucioMultiAccount?: string` to
`AuthViewModel`, this is a comma-separated string of the possible
account names.

Uses `react-modal` addon. In order to be able to style the modal and the
overlay (overlay ~= the container into which the modal is placed, ie the
container it is hovering in) using tailwind, the regex to discover
tailwind classes was adapted.