Components: Replaced button with `Link`
a button is supposed to change the page from DID List to the
corresponding DID Page. In order to use proper NextJS routing, I have
styled the `Link` component (which itself wraps the HTML `a` tag).

The new link will route to the proper page as expected.

Made tests run ok