Components: Added loading cards
and implemented them for `subscription page`. i opted not to use
shimmering or pulsating effects and went for the more informative
"loading...".

**NOTE**: added a new status to `BaseViewModel`, the status `pending`,
which is different from `success` and `error`.

The way this is supposed to work is that the nextjs page will initialise
the viewmodel that gets fed into the page as `{status: "pending"} as
<ViewModel>`. This viewmodel gets replaced once the static fetch is
completed.

If the ViewModel is in status `pending`, the page will render a loading
placeholder. this is a low-tech solution that does not rely on
`Suspense` etc from the react or nextjs frameworks, instead using the
ViewModels that we use natively.

**TODO**: repeat this for fetches that have an error later.