add middleware to redirect to login #280
The limitation with the middleware is that only GET requests
can be properly redirected after a successful login
Other types of requests cannot be cached in the session as the session
object in the middleware is not the same as the one made available
to the rest of the application. This is a bug with IronSession.