Components: Added NormalTable (nonstreamed) for PageDIDMetadata
PageDIDMetadata used to be implemented using a streamedtable, because we
worked under the assumption that the metadata would come in a stream.
now that it comes as a static fetch, we still want to keep the design of
the streamedtables. To solve this issue, we developed the `NormalTable`,
which is a streamedtable without the streaming functionalities. it still
keeps the design and the functionality provided by tanstack.

Also changed the fixtures to work properly.

Data comes in from the backend as a single viewmodel, which contains a
parameter `data: DIDKeyValuePair[]`, which contains the individual rows
of the tanstack table. This is unpacked and then used in the table.

**TODO** properly couple the `normal` and `streamed` table versions so
they are not developed completely independently. style changes in one
should also be reflected in the other!